Upgrade rails to 5.2.6
Context:

Atransitive dependency of rails was yanked (mimemagic 0.3.5) due to MIT vs GPL
licensing conflicts. The next version up (0.3.7) changed the license to GPL2
and now requires an external dependency (package `shared-mime-info`) to in order
to reference the freedesktop.org mime xml database.

Unfortunately, the installation of this gem generates a hardcoded path to the
mime xml database at build time, and this path must resolve at runtime as well.
As GoCD is installable across different platforms and distributions, this will
not work and would require users to install the database file at a specific path
on the host running GoCD server. This is a non-starter.

The most viable option to address this new dependency is to remove the dependency
altogether; upgrading rails to 5.2.5 or better solves this as the dependency on
this package (by way of marcel 0.3.x) has been severed since the update to marcel
1.0.x used by activemailer and activestorage as of 5.2.5 and above.

We should upgrade anyway to get the latest security patches and bugfixes on 5.x.