Add hard-dependency on generic virtual package
Not sure if this is a good idea; but is clear you need some type of runtime, while allowing that to be temurin or something else.
On RHEL-based distros this will correctly locate via /etc/alternatives, whereas on Debian-based distros this will require the user
to find it and update config as the alternatives symlinks are not version-specific.